requirements.txt
imageio==2.9.0
numpy==1.21.2
Pillow==8.3.2
